feat: properly handle session update
There's a tricky problem with Auth.js where the session does not get
updated with the latest data from the database. I can't tell if this is
an issue specific to database adapter + JWT session setups, but I didn't
have time to debug.

To solve this, I resorted to a hack where you query the database and
update the token manually inside the JWT callback. Not really ideal, but
it doesn't seem like there's any other way.