feat: implement authentication (#11)
* feat: send cookies and secrets in the trpc client

* feat: initial work on Google + Resend (e-mail) authentication

* feat: add Discord, GitHub and Google as OAuth providers

* refactor: unify login and signup page

In the world of OAuth and Magic Links, both the act of signing up and
logging in involves identical APIs. But we will still visually
distinguish them in the UI for the sake of user familiarity.

Note how things like names, usernames, etc. are not a part of the
initial registration process. We will redirect the user to a form to
complete this information if they are still missing from the database
(implement this later).

* feat: make the flow pages more responsive

* build: add tailwind formatter

* feat: add spinner component

* feat: provide global session context

* style: reformat

* feat: redirect if user is not logged in

* feat: initial work on post-registration flow

* feat: add registration form

* feat: properly handle session update

There's a tricky problem with Auth.js where the session does not get
updated with the latest data from the database. I can't tell if this is
an issue specific to database adapter + JWT session setups, but I didn't
have time to debug.

To solve this, I resorted to a hack where you query the database and
update the token manually inside the JWT callback. Not really ideal, but
it doesn't seem like there's any other way.

* feat: improve loading for front page

* feat: add log out button

* fix: remove modal popup

* chore: bump @binar-semicolon/api