fix: production bugs 💀
* build: relax template string restrictions

* fix: repair API communications for production

It looks like that tRPC is not working in production, for various
bizzare reasons. So here are my thoughts and hacks.

For client-side API calls, we will use Next.js rewrites to make it
easier to pass cross-origin cookies. According to this SE post:
https://stackoverflow.com/a/71902594/10625876, cookies can only be
passed among the same domain (but different ports) even with configured
CORS. It's unsourced and rather suspect, but it seems to fit my
experience. I just can't figure out a way to make cross-origin cookies
work during production.

For server-side API calls, we will call the API's domain directly.
I have actually tried three ways of doing this:
- Call $VERCEL_URL
- Call the actual public production domain
- Call the API's domain directly
But only the latter two works. The first one somehow returns an HTML
document. Again, I can't figure out why. But it might be a
Vercel-specific quirk.

* feat: include next.config.mjs in tsconfig.json

* chore: bump @binar-semicolon/api